Add setting for excluding files from autocompletion
I've had a problem with Sublime freezing for a long time when editing a certain .sql file (a file containing the thousands of \i commands needed to rebuild the database in question). The length of the file does not appear to be the deciding factor, which is why I made this an exclusion list rather than a max-file-length-for-autocompletion setting.